+++ packages/ooxml-oracle/example/index.js
@@ -1,4 +1,4 @@
-import { childrenOf } from '@superdoc-dev/ooxml-inspector';
+import { childrenOf } from '@superdoc-dev/ooxml-oracle';
 
 const children = childrenOf('w:p');
 console.debug('Children of w:p', children);