+++ packages/super-editor/src/core/super-converter/v2/handlers/registry/handler-registry.js
@@ -1,5 +1,5 @@
 import { registeredHandlers } from '../index.js';
-import { childrenOf } from '@superdoc-dev/ooxml-inspector';
+import { childrenOf } from '@superdoc-dev/ooxml-oracle';
 import {
   createCoverageGroupTitle,
   startCoverageGroup,